+++ src/main/java/org/uma/evolver/example/training/NSGAIIOptimizingNSGAIIForProblemZDT4.java
@@ -47,7 +47,7 @@ public class NSGAIIOptimizingNSGAIIForProblemZDT4 {
   private static final int PLOT_UPDATE_FREQUENCY = 1;
 
   public static void main(String[] args) throws IOException {
-    String yamlParameterSpaceFile = "NSGAIIDoubleFull.yaml";
+    String yamlParameterSpaceFile = "NSGAIIDouble.yaml";
 
     // Step 1: Select the target problem
     List<Problem<DoubleSolution>> trainingSet = List.of(new ZDT4());