+++ src/main/java/org/uma/evolver/example/validation/ZDTDTLZWFG2DStudy.java
@@ -45,7 +45,7 @@ public class ZDTDTLZWFG2DStudy {
   private static final int INDEPENDENT_RUNS = 25;
   private static final int MAX_EVALUATIONS = 20000;
   private static final int POPULATION_SIZE = 100;
-  private static final String YAML_FILE = "NSGAIIDoubleFull.yaml";
+  private static final String YAML_FILE = "NSGAIIDouble.yaml";
 
   public static void main(String[] args) throws IOException {
     String experimentBaseDirectory = "results/swevo/experiments"; // Dedicated folder